+++ src/instances/garbage/GarbageApp.vue
@@ -1,7 +1,0 @@
-<script setup lang="ts">
-import GarbageApp from '../../../../apps/garbage/src/app/GarbageApp.vue'
-</script>
-
-<template>
-  <GarbageApp />
-</template>