+++ src/instances/garbage/composables/useNewsletterSnapshot.ts
@@ -1,1 +1,0 @@
-export * from '../../../../../apps/garbage/src/app/composables/useNewsletterSnapshot.ts'