+++ src/instances/garbage/composables/useNewsletterView.ts
@@ -1,1 +1,0 @@
-export * from '../../../../../apps/garbage/src/app/composables/useNewsletterView.ts'