+++ src/instances/garbage/instance.ts
@@ -1,1 +1,0 @@
-export { workbenchInstanceRegistration } from '../../../../apps/garbage/src/instance-registration.ts'